+++ app/src/main/java/hr/avrbanac/docsis/app/PreEqualizationApplication.java
@@ -8,7 +8,7 @@ import javafx.stage.Stage;
 public class PreEqualizationApplication extends Application {
 
     @Override
-    public void start(Stage stage) throws Exception {
+    public void start(final Stage stage) throws Exception {
         FXMLLoader fxmlLoader = new FXMLLoader(PreEqualizationApplication.class.getResource("/views/preeq-view.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 1024, 768);
         stage.setTitle("PreEqualization");