+++ app/src/main/java/module-info.java
@@ -1,7 +1,0 @@
-module hr.avrbanac.preeequalization.app {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-    opens hr.avrbanac.docsis.app to javafx.fxml;
-    exports hr.avrbanac.docsis.app;
-}